pyproject.toml
[tool.poetry]
name = "netdevops"
version = "0.1.0"
description = ""
authors = ["Adam Angell <Adam.Angell@networkingb2a.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
xmltodict = "^0.13.0"
pydantic = "^2.9.1"
pydantic-yaml = "^1.3.0"
sqlalchemy = "^2.0.36"
pynetbox = "^7.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

